api/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
Pillow
torchvision
python-dotenv